requirements.txt
PyPDF2
python-dotenv
azure-core
azure-ai-documentintelligence
pdf2image
tenacity
azure-search-documents
azure-search
ipython
openai
azure-storage-blob
ipykernel
pandas
streamlit